Stripping out unnecessary kernels code
