--- kernels-run/test.py
@@ -1,1 +1,0 @@
-from util.env import RESULT_FILE_NATIVE
